+++ tests/pure/es.string.replace-all.js
@@ -26,6 +26,10 @@ QUnit.test('String#replaceAll', assert => {
   assert.same(replaceAll('aba', searcher, 'c'), 'foo');
   assert.same(replaceAll('aba', 'b'), 'aundefineda');
   assert.same(replaceAll('xxx', '', '_'), '_x_x_x_');
+  assert.same(replaceAll('121314', '1', '$$'), '$2$3$4', '$$');
+  assert.same(replaceAll('121314', '1', '$&'), '121314', '$&');
+  assert.same(replaceAll('121314', '1', '$`'), '212312134', '$`');
+  assert.same(replaceAll('121314', '1', '$\''), '213142314344', '$\'');
   if (STRICT) {
     assert.throws(() => replaceAll(null, 'a', 'b'), TypeError);
     assert.throws(() => replaceAll(undefined, 'a', 'b'), TypeError);